requirements.txt
fastapi
uvicorn[standard]
openai
faiss-cpu
numpy
pydantic
